FE: fix UI common
