+++ src/constants/router.constants.js
@@ -7,6 +7,7 @@ export const NamespaceRouter = {
   MY_PAGE: "My-page",
   LIST_CATEGORY: "List-category",
   PROFILE: "Profile",
+  SETTING_PROFILE: "Setting-profile",
   ALBUM_LIST: "Album-list",
   BOOKMARK_LIST: "Bookmark-list",
   NOT_FOUND: "Not-found",
