+++ src/views/my-page/constants/my-page.paths.js
@@ -1,6 +1,7 @@
 export const MyPagePaths = {
   MY_PAGE: "/my-page",
-  PROFILE: "/profile",
+  PROFILE: "/@:user_id",
+  SETTING_PROFILE: "/setting-profile",
   ALBUM_LIST: "/album-list",
   BOOKMARK_LIST: "/bookmark-list",
 };
