+++ src/views/my-page/screens/BookmarkList/BookmarkList.vue
@@ -1,54 +1,57 @@
 <template>
-  <div class="max-w-3xl mx-auto mb-10 min-h-[calc(100vh-64px-191px)] mt-7">
-    <return-to :to="MyPagePaths.MY_PAGE" />
+  <div class="mx-6">
+    <div class="max-w-5xl mx-auto mb-10 min-h-[calc(100vh-64px-191px)] mt-7">
+      <return-to :to="MyPagePaths.MY_PAGE" />
 
-    <h1 class="border-b py-2 text-lg font-semibold border-[#eee] mt-3 mb-7">
-      Albums đã lưu
-    </h1>
+      <h1 class="border-b py-2 text-lg font-semibold border-[#eee] mt-3 mb-7">
+        Albums đã lưu
+      </h1>
 
-    <button
-      :class="[
-        'font-sans text-[13px] border px-2 py-[3px] rounded-3xl border-[#ddd]  mr-3',
-        sort.value === initParams.sort ? 'bg-main text-white' : 'bg-gray',
-      ]"
-      v-for="(sort, index) in LIST_SORT"
-      :key="index"
-      @click="() => handlerSortAlbums(sort)"
-    >
-      {{ sort.label }}
-    </button>
+      <button
+        :class="[
+          'font-sans text-[13px] border px-2 py-[3px] rounded-3xl border-[#ddd]  mr-3',
+          sort.value === initParams.sort ? 'bg-main text-white' : 'bg-gray',
+        ]"
+        v-for="(sort, index) in LIST_SORT"
+        :key="index"
+        @click="() => handlerSortAlbums(sort)"
+      >
+        {{ sort.label }}
+      </button>
 
-    <div class="mt-5" v-if="isLoading">Loading...</div>
-    <div
-      v-else-if="
-        myPageStore?.listBookmarks && myPageStore?.listBookmarks?.length > 0
-      "
-      ref="masonry"
-    >
-      <div class="mt-5 gap-4 flex flex-wrap">
-        <item-album
-          :album="album"
-          v-for="album in myPageStore?.listBookmarks"
-          :key="album.id"
-          class="item-album !mb-0 h-full"
-        />
-      </div>
-
+      <div class="mt-5" v-if="isLoading">Loading...</div>
       <div
-        id="observe-visibility"
-        v-observe-visibility="{
-          callback: visibilityChanged,
-          throttle: 100,
-        }"
-      />
+        v-else-if="
+          myPageStore?.listBookmarks && myPageStore?.listBookmarks?.length > 0
+        "
+        ref="masonry"
+      >
+        <div class="mt-5 gap-4 flex flex-wrap">
+          <item-album
+            :album="album"
+            v-for="album in myPageStore?.listBookmarks"
+            :key="album.id"
+            class="item-album !mb-0 h-full"
+            overlay
+          />
+        </div>
 
-      <div class="flex justify-center mt-5" v-if="isLoadingScroll">
         <div
-          class="inline-block h-7 w-7 animate-spin rounded-full border-[3px] border-solid border-main border-current border-r-transparent align-[-0.125em] motion-reduce:animate-[spin_1.5s_linear_infinite]"
+          id="observe-visibility"
+          v-observe-visibility="{
+            callback: visibilityChanged,
+            throttle: 100,
+          }"
         />
+
+        <div class="flex justify-center mt-5" v-if="isLoadingScroll">
+          <div
+            class="inline-block h-7 w-7 animate-spin rounded-full border-[3px] border-solid border-main border-current border-r-transparent align-[-0.125em] motion-reduce:animate-[spin_1.5s_linear_infinite]"
+          />
+        </div>
       </div>
+      <div v-else>not data</div>
     </div>
-    <div v-else>not data</div>
   </div>
 </template>
 
@@ -123,13 +126,4 @@ onUnmounted(() => {
 });
 </script>
 
-<style lang="css" scoped>
-.item-album {
-  flex: 1 0 auto;
-  height: 180px;
-}
-
-.item-album:last-child {
-  max-width: 400px;
-}
-</style>
+<style lang="css"></style>
